# Conflicts:
#	server/src/main/java/com/runstory/api/controller/FeedController.java
#	server/src/main/java/com/runstory/api/controller/MainController.java
#	server/src/main/java/com/runstory/domain/feed/dto/FeedCommentDto.java
#	server/src/main/java/com/runstory/domain/feed/dto/FeedRecommentDto.java
#	server/src/main/java/com/runstory/domain/feed/entity/FeedComment.java
#	server/src/main/java/com/runstory/domain/feed/entity/FeedRecomment.java


Former-commit-id: e8e8433c0bb0f8e722dcdff48aa4ab248f7ce65e

+++ server/src/main/java/com/runstory/service/FeedService.java
@@ -1,15 +1,18 @@
 package com.runstory.service;
 
+import com.runstory.api.request.FeedCommentReqDto;
 import com.runstory.api.request.FeedReqDto;
 
 import com.runstory.api.request.FeedReqDto;
 import com.runstory.api.response.FeedResDto;
+import com.runstory.api.response.SimpleFeedResDto;
+import com.runstory.common.util.FileUtil;
 import com.runstory.domain.feed.PublicScope;
+import com.runstory.domain.feed.dto.FeedCommentDto;
 import com.runstory.domain.feed.dto.FeedDto;
-import com.runstory.domain.feed.entity.Feed;
-import com.runstory.domain.feed.entity.FeedFile;
-import com.runstory.domain.feed.entity.FeedLike;
+import com.runstory.domain.feed.entity.*;
 import com.runstory.domain.hashtag.HashtagType;
+import com.runstory.domain.hashtag.dto.HashtagDto;
 import com.runstory.domain.hashtag.entity.Hashtag;
 import com.runstory.domain.hashtag.entity.SelectedHashtag;
 import com.runstory.domain.user.entity.Follow;
@@ -28,14 +31,12 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.net.InetAddress;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -50,6 +51,8 @@ public class FeedService {
     private final  HashtagRepository hashtagRepository;
     private final FeedFileRepository feedFileRepository;
     private final FeedLikeRepository feedLikeRepository;
+    private final FeedCommentRepository feedCommentRepository;
+    private final FeedReCommentRepository feedReCommentRepository;
 
     public List<FeedDto> findAll(){
         List<Feed> feeds = feedRepository.findAll();
@@ -101,7 +104,6 @@ public class FeedService {
         List<FeedResDto> result = new ArrayList<>();
         for(Feed f :feeds){
             FeedLike feedLike = feedLikeRepository.findByFeedIdAndUserId(f.getFeedId(),user.getUserSeq());
-            System.out.println(feedLike==null?false:true);
             FeedResDto feedResDto = new FeedResDto(f, feedLike);
             result.add(feedResDto);
         }
@@ -138,6 +140,12 @@ public class FeedService {
         return feedRepository.findByFeedIdLessThanAndPublicScopeNotInAndUserInOrderByFeedIdDesc(lastFeedId, scope, followers, pageRequest); // JPA 쿼리 메소드
     }
 
+    public FeedResDto findByFeedId(Long userId, Long feedId){
+        Feed feed= feedRepository.findByFeedId(feedId);
+        FeedLike feedLike = feedLikeRepository.findByFeedIdAndUserId(feedId, userId);
+        FeedResDto result = new FeedResDto(feed, feedLike);
+        return result;
+    }
     @Transactional
     public Feed saveFeed(FeedReqDto feedReqDto, MultipartFile [] files) throws IOException {
         User user = userRepository.findByUserSeq(feedReqDto.getUserId());
@@ -151,13 +159,11 @@ public class FeedService {
     }
     @Transactional
     public void saveFiles(Feed feed, MultipartFile [] files) throws IOException {
-        for(MultipartFile file: files){
-            String fileName = file.getOriginalFilename();
-            String filePath = "C:/runTogether/uploads/feed/" + UUID.randomUUID()+fileName;
-            Path path = Paths.get(filePath);
-            Files.write(path, file.getBytes());
-
-            FeedFile feedFile = new FeedFile(feed, fileName,filePath);
+        for(MultipartFile feedfile: files){
+            String hostname = InetAddress.getLocalHost().getHostName();
+            FileUtil fileUtil = new FileUtil();
+            HashMap<String, String> file = fileUtil.fileCreate(hostname, "feeds",feedfile);
+            FeedFile feedFile = new FeedFile(feed, file.get("filename"), file.get("filepath"));
             feedFileRepository.save(feedFile);
         }
     }
@@ -170,40 +176,51 @@ public class FeedService {
         }
     }
 
+    /**
+     * 피드 업데이트는 피드의 내용, 해시태그가 수정 가능하다.(***추후 파일 수정도 가능하게 만들지 고려)
+     * @param feed(FeedReqDto)
+     * @param feedId
+     * @return
+     * @throws IOException
+     */
     @Transactional
     public Feed updateFeed(FeedReqDto feed, Long feedId) throws IOException {
         //피드 가져오기
-        Feed f = feedRepository.findByFeedId(feedId);
-        //내용 수정
-        f.setContent(feed.getContent());
-        f.setPublicScope(feed.getPublicScope());
+        Feed f = feedRepository.findByFeedIdAndUserUserSeq(feedId, feed.getUserId());
+        if(f!=null) {
+            //내용 수정
+            f.setContent(feed.getContent());
+            f.setPublicScope(feed.getPublicScope());
 
-        //해시태그 수정
-        List<Long> tags = feed.getSelectedHashTags();
-        Collections.sort(tags);
-        List<SelectedHashtag> selectedHashtags = selectedHashtagRepository.findByFeedIdOrderByHashtagIdAsc(feedId);    //DB에 저장된 해시태그
+            //해시태그 수정
+            List<Long> tags = feed.getSelectedHashTags();
+            Collections.sort(tags);
+            List<SelectedHashtag> selectedHashtags = selectedHashtagRepository.findByFeedIdOrderByHashtagIdAsc(feedId);    //DB에 저장된 해시태그
 
-        //저장된 해시태그 개수 비교
-        if(tags.size()!=selectedHashtags.size()){
-            System.out.println("해시태그 개수 다름");
-            //selectedhashtag 삭제 후 저장
-            selectedHashtagRepository.deleteSelectedHashtagByFeedId(feedId);
-            saveHashtags(f,feed.getSelectedHashTags());
-        } else {
-            //개수가 같으면 하나하나 비교
-            for(int i =0;i<tags.size();i++){
-                if(tags.get(i)!=selectedHashtags.get(i).getHashtag().getHashtagId()){
-                    System.out.println("해시태그 다시 저장");
-                    //selectedhashtag 삭제 후 저장
-                    selectedHashtagRepository.deleteSelectedHashtagByFeedId(feedId);
-                    saveHashtags(f,feed.getSelectedHashTags());
-                    break;
+            //저장된 해시태그 개수 비교
+            if (tags.size() != selectedHashtags.size()) {
+                System.out.println("해시태그 개수 다름");
+                //selectedhashtag 삭제 후 저장
+                selectedHashtagRepository.deleteSelectedHashtagByFeedId(feedId);
+                saveHashtags(f, feed.getSelectedHashTags());
+            } else {
+                //개수가 같으면 하나하나 비교
+                for (int i = 0; i < tags.size(); i++) {
+                    if (tags.get(i) != selectedHashtags.get(i).getHashtag().getHashtagId()) {
+                        System.out.println("해시태그 다시 저장");
+                        //selectedhashtag 삭제 후 저장
+                        selectedHashtagRepository.deleteSelectedHashtagByFeedId(feedId);
+                        saveHashtags(f, feed.getSelectedHashTags());
+                        break;
+                    }
                 }
             }
+            f.setSelectedHashtags(selectedHashtagRepository.findByFeedIdOrderByHashtagIdAsc(feedId));
+            Feed result = feedRepository.save(f);
+
+            return result;
         }
-        f.setSelectedHashtags(selectedHashtagRepository.findByFeedIdOrderByHashtagIdAsc(feedId));
-        Feed result = feedRepository.save(f);
-        return result;
+        return null;
     }
 
     @Transactional
@@ -213,5 +230,120 @@ public class FeedService {
             return true;
         }
         return false;
+    }
+
+    /**
+     * 검색페이지 해시태그 기반 피드 검색
+     * @param hashtagId
+     * @param lastFeedId
+     * @param size
+     * @return
+     */
+    public List<SimpleFeedResDto> searchByHashtag(Long hashtagId, Long lastFeedId, int size){
+        PageRequest pageRequest = PageRequest.of(0, size);
+        //해시태그에 맞는 피드 아이디 리스트를 가져온다.
+        List<SelectedHashtag> selectedHashtags = selectedHashtagRepository.findByHashtag_HashtagIdAndFeedNotNull(hashtagId);
+        List<Long> feedIds = selectedHashtags.stream().map(s->s.getFeed().getFeedId()).collect(Collectors.toList());
+        List<PublicScope> scope = new ArrayList<>();
+        scope.add(PublicScope.PUBLIC);
+        //검색한 해시태그에 맞는 피드들을 조회한다.
+        Page<Feed> feeds = feedRepository.findByFeedIdLessThanAndFeedIdInAndPublicScopeInOrderByFeedIdDesc
+                (lastFeedId,feedIds, scope, pageRequest);
+        List<FeedDto> tmp = feeds.stream().map(f->new FeedDto(f)).collect(Collectors.toList());
+        List<SimpleFeedResDto> result = tmp.stream().map(t->new SimpleFeedResDto(t)).collect(Collectors.toList());
+        return result;
+    }
+
+    /**
+     * 피드 좋아요 저장
+     * @param feedId
+     * @param userId
+     * @return
+     */
+    @Transactional
+    public FeedLike saveFeedLiKe(Long feedId, Long userId){
+        Feed feed = feedRepository.findByFeedId(feedId);
+        User user = userRepository.findByUserSeq(userId);
+        FeedLike feedLike = new FeedLike(feed, user);
+        return feedLikeRepository.save(feedLike);
+    }
+
+    @Transactional
+    public void deleteFeedLike(Long feedId, Long userId){
+        FeedLike feedLike = feedLikeRepository.findByFeedIdAndUserId(feedId, userId);
+        feedLikeRepository.deleteById(feedLike.getFeedLikeId());
+    }
+
+    // Feed 댓글 생성
+    @Transactional
+    public Long createFeedComment(FeedCommentReqDto feedCommentReqDto, Long userSeq){
+        Feed feed = feedRepository.findByFeedId(feedCommentReqDto.getId());
+        User user = userRepository.findByUserSeq(userSeq);
+        FeedComment feedComment = FeedComment.builder()
+                .feed(feed)
+                .user(user)
+                .content(feedCommentReqDto.getContent())
+                .build();
+        feedCommentRepository.save(feedComment);
+        return feedComment.getFeedCommentId();
+    }
+
+    // Feed 댓글 삭제
+    @Transactional
+    public Long deleteFeedComment(Long commentId, Long userSeq){
+        User user = userRepository.findByUserSeq(userSeq);
+        FeedComment feedComment = feedCommentRepository.findByFeedCommentIdAndUser(commentId, user);
+        if (feedComment == null){
+                return null;
+            }else{ // 만약 있으면
+                feedCommentRepository.deleteById(commentId);
+                return commentId;
+            }
+    }
+
+    @Transactional
+    public Long createFeedRecomment(FeedCommentReqDto feedCommentReqDto, Long userSeq){
+        FeedComment feedComment = feedCommentRepository.findByFeedCommentId(feedCommentReqDto.getId());
+        User user = userRepository.findByUserSeq(userSeq);
+        FeedRecomment feedRecomment = FeedRecomment.builder()
+                .feedComment(feedComment)
+                .user(user)
+                .cotent(feedCommentReqDto.getContent())
+                .build();
+        feedReCommentRepository.save(feedRecomment);
+        return feedRecomment.getFeedRecommnetId();
+    }
+
+    @Transactional
+    public Long deleteFeedReComment(Long recommentId, Long userSeq){
+        User user = userRepository.findByUserSeq(userSeq);
+        FeedRecomment feedRecomment = feedReCommentRepository.findByFeedRecommnetIdAndUser(recommentId, user);
+        if (feedRecomment == null){
+            return -1L;
+        }else{ // 만약 있으면
+            feedReCommentRepository.deleteById(recommentId);
+            return recommentId;
+        }
+    }
+
+
+    public List<FeedCommentDto> getFeedDetail(Long feedId){
+        List<FeedCommentDto> result = new ArrayList<>();
+        Feed feed = feedRepository.findByFeedId(feedId);
+        for (FeedComment feedComment : feed.getFeedComments()){
+            FeedCommentDto feedCommentDto = new FeedCommentDto(feedComment);
+            result.add(feedCommentDto);
+        }
+        return result;
+    }
+
+    /**
+     * 해시태그 리스트 가져오기
+     * @return result
+     */
+    public List<HashtagDto> getHashtags(){
+        List<Hashtag> hashtags = hashtagRepository.findAll();
+        List<HashtagDto> result = hashtags.stream().map(h->new HashtagDto(h)).collect(Collectors.toList());
+        return result;
     }
 }
